marketing-multi-agents/sql-agent/pyproject.toml
[project]
name = "sql-agent"
version = "20250902.1"
description = ""
authors = [
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "google-adk (>=1.23.0,<1.25.1)",
    "google-genai (>=1.60.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "immutabledict (>=4.2.1,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "google-cloud-discoveryengine",
    "google-cloud-aiplatform[agent-engines] (>=1.126.1)",
    "google-cloud-bigquery (>=3.35.1,<4.0.0)",
    "regex (>=2024.5.15,<2025.0.0)",
    "sqlglot (>=25.3.0,<26.0.0)",
    "absl-py (>=2.3.1,<3.0.0)",
    "db-dtypes (>=1.4.3,<2.0.0)"
]

[tool.poetry]
packages = [
    { include = "sql_agent" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
